modify comment_delete

+++ reviews/views.py
@@ -5,6 +5,7 @@ from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from .models import Review, Comment
 from .forms import ReviewForm, CommentForm
+from django.http import HttpResponseForbidden
 import random
 
 # Create your views here.
@@ -94,10 +95,13 @@ def comment_create(request, pk):
         comment.save()
     return redirect("reviews:detail", review.pk)
 
+
+@login_required
 def comment_delete(request, comment_pk, pk): # 마지막에 특정 리뷰에 대한 pk가 필요함
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    return redirect("reviews:detail", pk)
+    if request.user.is_authenticated and request.user == comment.user:
+        comment.delete()
+        return redirect("reviews:detail", pk)
 
 @login_required
 def like(request, pk):
